requirements.txt
Django~=3.1.7
djangorestframework~=3.12.2
gpxpy~=1.3.5